# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eltuvchi API",
        "description": "Eltuvchi platformasi uchun API hujjatlari",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/dispatcher/drivers": {
            "get": {
                "tags": [
                    "Dispatcher Drivers"
                ],
                "summary": "Get list of drivers in your taxopark",
                "operationId": "19290bb9b730a32259013e65390587c1",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by driver status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "archived"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of drivers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DispatcherDriver"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dispatcher Drivers"
                ],
                "summary": "Create a new driver",
                "operationId": "d8a5ff95264b070dbf74fd022b6b40c9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "archived"
                                        ]
                                    },
                                    "balance": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "points": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "details": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "settings": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Driver created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DispatcherDriver"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/dispatcher/drivers/{driver}": {
            "get": {
                "tags": [
                    "Dispatcher Drivers"
                ],
                "summary": "Get single driver details",
                "operationId": "73d9692bcd7fe31a49f640430322013c",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Driver details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DispatcherDriver"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Dispatcher Drivers"
                ],
                "summary": "Update driver",
                "operationId": "ed62f404cb21ce6992e59347b107ff01",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "archived"
                                        ]
                                    },
                                    "balance": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "points": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "details": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "settings": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Driver updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DispatcherDriver"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dispatcher Drivers"
                ],
                "summary": "Deactivate driver",
                "operationId": "28afd5219d1f24f8104425821fab5d74",
                "parameters": [
                    {
                        "name": "driver",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Driver deactivated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DispatcherDriver"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DispatcherDriver": {
                "title": "Dispatcher Driver",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "taxopark_id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    },
                    "balance": {
                        "type": "number",
                        "format": "float"
                    },
                    "points": {
                        "type": "number",
                        "format": "float"
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "settings": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "type": "object",
                        "nullable": true
                    },
                    "taxopark": {
                        "type": "object",
                        "nullable": true
                    },
                    "orders": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DispatcherOrder"
                        }
                    }
                },
                "type": "object"
            },
            "DispatcherOrder": {
                "title": "Dispatcher Order",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "client_id": {
                        "type": "integer"
                    },
                    "driver_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "route_id": {
                        "type": "integer"
                    },
                    "passengers": {
                        "type": "integer"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "time": {
                        "type": "string"
                    },
                    "price_order": {
                        "type": "number",
                        "format": "float"
                    },
                    "client_deposit": {
                        "type": "number",
                        "format": "float"
                    },
                    "driver_payment": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "discount_percent": {
                        "type": "integer",
                        "nullable": true
                    },
                    "discount_summ": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string"
                    },
                    "optional_phone": {
                        "type": "string",
                        "nullable": true
                    },
                    "note": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum authentication using Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}